Remove the mt_size_t type based on a misunderstanding. size_t is supposed to be only for in-memory sizes. For on disk sizes off_t (and by extension mt_off_t) should be used. off_t is not just for offsets, but also for sizes. Read/write-like calls still take size_t arguments, because the amount of data to be read in one go not only needs to fit on disk, but also in memory
git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@665 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ floppyd.c
@@ -1236,7 +1236,7 @@ void serve_client(int sockhandle, const char *const*device_name,
 #endif
 
 				lseek(devFd,
-				      get_dword(parm, 0),
+				      (off_t) get_dword(parm, 0),
 				      (int) get_dword(parm, 4));
 				send_reply(devFd,
 					   sock,
